# api.yaml
{
  "openapi": "3.0.0",
  "paths": {
    "/api/user/{user_id}": {
      "get": {
        "operationId": "UserController_getUserById",
        "summary": "Get user profile information by ID",
        "description": "Fetches user profile information",
        "parameters": [
          {
            "name": "user_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User profile fetched successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - invalid or missing token"
          },
          "404": {
            "description": "User not found"
          }
        },
        "tags": [
          "user"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "operationId": "UserController_updateUser",
        "summary": "Upload user information",
        "description": "Updates the information of the user authenticated",
        "parameters": [
          {
            "name": "user_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - empty fields"
          },
          "401": {
            "description": "Unauthorized - token missing or invalid"
          },
          "403": {
            "description": "Forbidden - Unauthorized access"
          },
          "404": {
            "description": "User not found"
          }
        },
        "tags": [
          "user"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/user/{user_id}/avatar": {
      "post": {
        "operationId": "UserController_createAvatar",
        "summary": "Upload user avatar",
        "description": "Uploads an image file (.png, .jpeg, .webp) as an avatar",
        "parameters": [
          {
            "name": "user_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Avatar file in \"file\" field",
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/AvatarUploadDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Avatar uploaded successfully"
          },
          "400": {
            "description": "Bad request - invalid type or large file"
          },
          "401": {
            "description": "Unauthorized - invalid or missing token"
          },
          "403": {
            "description": "Forbidden - Unauthorized access to upload avatar"
          }
        },
        "tags": [
          "user"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "operationId": "UserController_updateAvatar",
        "summary": "Update user avatar",
        "parameters": [
          {
            "name": "user_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Avatar file in \"file\" field",
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/AvatarUploadDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Avatar updated"
          },
          "400": {
            "description": "Bad request"
          },
          "403": {
            "description": "Forbidden – not your own profile"
          }
        },
        "tags": [
          "user"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "UserController_deleteAvatar",
        "summary": "Delete user avatar",
        "parameters": [
          {
            "name": "user_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Avatar deleted"
          },
          "403": {
            "description": "Forbidden – not your own profile"
          },
          "404": {
            "description": "User not found"
          }
        },
        "tags": [
          "user"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "smart-home",
    "description": "Smart home application",
    "version": "1.0",
    "contact": {

    }
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "example": "uuid-123-456",
            "description": "Unique identifier for the user"
          },
          "email": {
            "type": "string",
            "example": "d13@test.com"
          },
          "firstName": {
            "type": "string",
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "example": "Doe"
          },
          "avatarUrl": {
            "type": "string",
            "example": "http://example.com/storage/avatar.png",
            "description": "URL to the user avatar image"
          }
        },
        "required": [
          "userId",
          "email",
          "firstName",
          "lastName",
          "avatarUrl"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "description": "First name of the user",
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "description": "Last name of the user",
            "example": "Doe"
          }
        }
      },
      "AvatarUploadDto": {
        "type": "object",
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "description": "Image file to upload"
          }
        },
        "required": [
          "file"
        ]
      },
...
    }
  }
}